+++ chain/thorchain/module_bank.go
@@ -2,7 +2,6 @@ package thorchain
 
 import (
 	"context"
-	"fmt"
 
 	sdkmath "cosmossdk.io/math"
 
@@ -12,27 +11,9 @@ import (
 	"github.com/strangelove-ventures/interchaintest/v8/ibc"
 )
 
-// BankSend sends tokens from one account to another.
-func (tn *ChainNode) BankSend(ctx context.Context, keyName string, amount ibc.WalletAmount) error {
-	_, err := tn.ExecTx(ctx,
-		keyName, "bank", "send", keyName,
-		amount.Address, fmt.Sprintf("%s%s", amount.Amount.String(), amount.Denom),
-	)
-	return err
-}
-
 // Deprecated: use BankSend instead
 func (tn *ChainNode) SendFunds(ctx context.Context, keyName string, amount ibc.WalletAmount) error {
 	return tn.BankSend(ctx, keyName, amount)
-}
-
-// BankMultiSend sends an amount of token from one account to multiple accounts.
-func (tn *ChainNode) BankMultiSend(ctx context.Context, keyName string, addresses []string, amount sdkmath.Int, denom string) error {
-	cmd := append([]string{"bank", "multi-send", keyName}, addresses...)
-	cmd = append(cmd, fmt.Sprintf("%s%s", amount, denom))
-
-	_, err := tn.ExecTx(ctx, keyName, cmd...)
-	return err
 }
 
 // GetBalance fetches the current balance for a specific account address and denom.
