+++ chain/thorchain/module_thorchain.go
@@ -1,0 +1,17 @@
+package thorchain
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/strangelove-ventures/interchaintest/v8/ibc"
+)
+
+// BankSend sends tokens from one account to another.
+func (tn *ChainNode) BankSend(ctx context.Context, keyName string, amount ibc.WalletAmount) error {
+	_, err := tn.ExecTx(ctx,
+		keyName, "thorchain", "send",
+		amount.Address, fmt.Sprintf("%s%s", amount.Amount.String(), amount.Denom),
+	)
+	return err
+}
